Relocate new default template yaml files
