+++ packages/cli/templates/default-backend-module.yaml
@@ -1,6 +1,6 @@
 # lib/new/factories/backendModule.ts
 description: A new backend module that extends an existing backend plugin with additional features
-template: ../default-backend-module
+template: ./default-backend-module
 targetPath: plugins
 prompts:
   - id: test
