+++ packages/cli/templates/default-backend-plugin.yaml
@@ -1,6 +1,6 @@
 # lib/new/factories/backendPlugin.ts
 description: A new backend plugin
-template: ../default-backend-plugin
+template: ./default-backend-plugin
 targetPath: plugins
 prompts:
   - id: test
