+++ packages/cli/templates/default-common-plugin-package.yaml
@@ -1,6 +1,6 @@
 # lib/new/factories/pluginCommon.ts
 description: A new isomorphic common plugin package
-template: ../default-common-plugin-package
+template: ./default-common-plugin-package
 targetPath: plugins
 prompts:
   - id: test
