+++ packages/cli/templates/default-node-plugin-package.yaml
@@ -1,6 +1,6 @@
 # lib/new/factories/pluginNode.ts
 description: A new Node.js library plugin package
-template: ../default-node-plugin-package
+template: ./default-node-plugin-package
 targetPath: plugins
 prompts:
   - id: test
