+++ packages/cli/templates/default-plugin.yaml
@@ -1,6 +1,6 @@
 # lib/new/factories/frontendPlugin.ts
 description: A new frontend plugin
-template: ../default-plugin
+template: ./default-plugin
 targetPath: plugins
 prompts:
   - id: test
