+++ packages/cli/templates/index.ts
