+++ packages/cli/templates/scaffolder-module.yaml
@@ -1,6 +1,6 @@
 # lib/new/factories/scaffolderModule.ts
 description: A module exporting custom actions for @backstage/plugin-scaffolder-backend
-template: ../scaffolder-module
+template: ./scaffolder-module
 targetPath: plugins
 prompts:
   - id: test
